feature: some turno endpoints and services and modified ICrudService interface

+++ src/main/java/BackendC3/ClinicaOdontologica/service/ITurnoService.java
@@ -1,7 +1,6 @@
 package BackendC3.ClinicaOdontologica.service;
 
-import BackendC3.ClinicaOdontologica.entity.Turno;
-import BackendC3.ClinicaOdontologica.entity.TurnoDto;
+import BackendC3.ClinicaOdontologica.dto.TurnoDto;
 
-public interface ITurnoService extends ICrudService<TurnoDto, Integer>{
+public interface ITurnoService extends ICrudService<TurnoDto, Long>{
 }
